Move definitions of Operator and LinearOperator base classes
